metal: enable wave64 mat-mul for MoE experts (2x MoE prefill)
Fixes the correctness bug that had kernel_mul_mm_id_w64 disabled, and turns it on.

The fault was not in the tiling, the id map, or the dispatch -- all of which were
correct. It was a code-generation bug in AMD's Metal compiler: multiplying a uint64
stride by a *short* index produces a wrong result, with the short landing in the high
word. So args.nb12*i12 for i12=128 evaluated to nb12<<32 (4398046511104 instead of
131072), the B pointer landed far out of bounds, and the tile computed from
uninitialized memory.

That is why the failure looked so strange: it appeared at exactly token 128 and
beyond, independent of m, n_experts, and tile geometry, because 128 is where a short
index first pushes the miscompiled product past anything the buffer could cover.
Casting the indices to int fixes it; the dense kernel was never affected because it
already used int throughout.

Isolated with a standalone reproducer that runs the same graph on CPU and Metal and
diffs per column. The decisive measurement compared four addressing forms in-kernel:

    args.nb12 * i12          (short)  -> nan
    args.nb12 * (int)i12              -> correct
    (ulong)args.nb12 * (ulong)128     -> correct

Results on a Radeon Pro Vega II, Qwen3-30B-A3B-Q4_K_M:

    pp512    88.4 -> 179.2 t/s   (2.03x)
    pp2048   76.4 -> 172.2 t/s   (2.25x)
    tg32     46.3 ->  46.4 t/s   (unchanged, as expected)

Correctness: MUL_MAT_ID now passes 1087/1087. Perplexity on wikitext-2 (80 chunks) is
9.6627 against 9.6700 for the previous mat-vec path and 9.6356 for CPU -- slightly
closer to the CPU reference than before. Dense mat-mul is unchanged at 162.9 t/s.

+++ ggml/src/ggml-metal/ggml-metal.metal
@@ -10947,9 +10947,13 @@ kernel void kernel_mul_mm_id_w64(
 
             device const T1 * y = nullptr;
             if (col_ok) {
-                const int   id  = ids_i32[im*args.ne21 + r1 + n];
-                const short i11 = (id % args.ne20) % args.ne11;
-                const short i12 = (id / args.ne20);
+                const int id = ids_i32[im*args.ne21 + r1 + n];
+
+                // note: these MUST be int, not short. Multiplying a uint64 stride by a short
+                // is miscompiled on AMD's Metal compiler - the short lands in the high word,
+                // so nb12*128 comes out as nb12<<32 and the read lands far out of bounds.
+                const int i11 = (id % args.ne20) % args.ne11;
+                const int i12 = (id / args.ne20);
 
                 y = (device const T1 *)(src1 + args.nb11*i11 + args.nb12*i12);
             }
@@ -10982,9 +10986,9 @@ kernel void kernel_mul_mm_id_w64(
         const int col = r1 + tn + j;
         if (col >= neh1) continue;
 
-        const int   id  = ids_i32[im*args.ne21 + col];
-        const short ide = id % args.ne20;
-        const short idt = id / args.ne20;
+        const int id  = ids_i32[im*args.ne21 + col];
+        const int ide = id % args.ne20;   // int, not short - see the note in the B load above
+        const int idt = id / args.ne20;
 
         device float * D = (device float *) dst + ide*args.ne0 + idt*(size_t)args.ne1*args.ne0;
